requirements.txt
streamlit>=1.21.0
pandas>=1.5.3
numpy>=1.24.2
matplotlib>=3.7.1
seaborn>=0.12.2
python-dotenv>=1.0.0
requests>=2.32.3
